Bug 2265: Use streaming for DeleteData message
The WriteData and MergeData messages were changed to use streaming. We
might as well do it for DeleteData as well as the YangInstanceIdentifier
streaming is faster.

+++ opendaylight/md-sal/sal-distributed-datastore/src/main/java/org/opendaylight/controller/cluster/datastore/messages/DeleteData.java
@@ -8,15 +8,26 @@
 
 package org.opendaylight.controller.cluster.datastore.messages;
 
+import java.io.Externalizable;
+import java.io.IOException;
+import java.io.ObjectInput;
+import java.io.ObjectOutput;
+import org.opendaylight.controller.cluster.datastore.DataStoreVersions;
 import org.opendaylight.controller.cluster.datastore.util.InstanceIdentifierUtils;
+import org.opendaylight.controller.cluster.datastore.utils.SerializationUtils;
 import org.opendaylight.controller.protobuff.messages.transaction.ShardTransactionMessages;
 import org.opendaylight.yangtools.yang.data.api.YangInstanceIdentifier;
 
-public class DeleteData implements SerializableMessage {
+public class DeleteData implements VersionedSerializableMessage, Externalizable {
+    private static final long serialVersionUID = 1L;
 
-    public static final Class<ShardTransactionMessages.DeleteData> SERIALIZABLE_CLASS = ShardTransactionMessages.DeleteData.class;
+    public static final Class<DeleteData> SERIALIZABLE_CLASS = DeleteData.class;
 
-    private final YangInstanceIdentifier path;
+    private YangInstanceIdentifier path;
+    private short version;
+
+    public DeleteData() {
+    }
 
     public DeleteData(final YangInstanceIdentifier path) {
         this.path = path;
@@ -26,13 +37,46 @@ public class DeleteData implements SerializableMessage {
         return path;
     }
 
-    @Override public Object toSerializable() {
-        return ShardTransactionMessages.DeleteData.newBuilder()
-            .setInstanceIdentifierPathArguments(InstanceIdentifierUtils.toSerializable(path)).build();
+    public short getVersion() {
+        return version;
     }
 
-    public static DeleteData fromSerializable(final Object serializable){
-        ShardTransactionMessages.DeleteData o = (ShardTransactionMessages.DeleteData) serializable;
-        return new DeleteData(InstanceIdentifierUtils.fromSerializable(o.getInstanceIdentifierPathArguments()));
+    @Override
+    public void readExternal(ObjectInput in) throws IOException, ClassNotFoundException {
+        version = in.readShort(); // Read the version - don't need to do anything with it now
+        path = SerializationUtils.deserializePath(in);
+    }
+
+    @Override
+    public void writeExternal(ObjectOutput out) throws IOException {
+        out.writeShort(version);
+        SerializationUtils.serializePath(path, out);
+    }
+
+    @Override
+    public Object toSerializable(short toVersion) {
+        if(toVersion >= DataStoreVersions.LITHIUM_VERSION) {
+            version = toVersion;
+            return this;
+        } else {
+            // To base or R1 Helium version
+            return ShardTransactionMessages.DeleteData.newBuilder().setInstanceIdentifierPathArguments(
+                    InstanceIdentifierUtils.toSerializable(path)).build();
+        }
+    }
+
+    public static DeleteData fromSerializable(final Object serializable) {
+        if(serializable instanceof DeleteData) {
+            return (DeleteData) serializable;
+        } else {
+            // From base or R1 Helium version
+            ShardTransactionMessages.DeleteData o = (ShardTransactionMessages.DeleteData) serializable;
+            return new DeleteData(InstanceIdentifierUtils.fromSerializable(o.getInstanceIdentifierPathArguments()));
+        }
+    }
+
+    public static boolean isSerializedType(Object message) {
+        return SERIALIZABLE_CLASS.isInstance(message) ||
+               message instanceof ShardTransactionMessages.DeleteData;
     }
 }
